implement edit category with actions
