+++ src/components/CategoryForm/CategoryForm.jsx
@@ -1,6 +1,14 @@
-import { Form, useActionData, useNavigate } from "react-router-dom";
+import {
+	Form,
+	useActionData,
+	useNavigate,
+	json,
+	redirect,
+} from "react-router-dom";
 
 import styles from "./CategoryForm.module.css";
+
+import { getAuthToken } from "../../../util/auth";
 
 const CategoryForm = ({ method, category }) => {
 	const data = useActionData();
@@ -13,7 +21,7 @@ const CategoryForm = ({ method, category }) => {
 	}
 
 	return (
-		<Form method="post" className={styles.form}>
+		<Form method={method} className={styles.form}>
 			{data && data.msg && (
 				<ul>
 					<li key={data.msg}>{data.msg}</li>
@@ -50,3 +58,40 @@ const CategoryForm = ({ method, category }) => {
 };
 
 export default CategoryForm;
+
+export const action = async ({ request, params }) => {
+	const method = request.method;
+
+	const data = await request.formData();
+
+	const eventData = {
+		name: data.get("title"),
+	};
+
+	const token = getAuthToken();
+
+	let url = "http://localhost:8080/api/categories";
+
+	if (method === "PUT") {
+		const eventId = params.categoryId;
+		url = "http://localhost:8080/api/categories/" + eventId;
+	}
+
+	const response = await fetch(url, {
+		method: method,
+		headers: {
+			"Content-Type": "application/json",
+			"x-token": token,
+		},
+		body: JSON.stringify(eventData),
+	});
+
+	if (!response.ok) {
+		// throw json({ msg: "Could not save category." }, { status: 500 });
+		return response;
+	}
+
+	console.log(await response);
+
+	return redirect("/categories");
+};
